requirements.txt
fastapi
minio
Pillow
python-dotenv
stability-sdk
uvicorn[standard]